Emulado escopo da fixture scope, module ou session close #46
